# Desktop/itm/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Road Requests API
  version: 1.0.0
  description: API for managing road repair requests

servers:
  - url: http://localhost:3001

paths:
  /requests:
    get:
      summary: Get all requests
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Request"

    post:
      summary: Create new request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCreate"
      responses:
        "201":
          description: Created

  /requests/{id}:
    get:
      summary: Get request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "404":
          description: Not found

    put:
      summary: Update request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUpdate"
      responses:
        "200":
          description: Updated

    delete:
      summary: Delete request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

components:
  schemas:
    Request:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        address:
          type: string
        status:
          type: string

    RequestCreate:
      type: object
      required:
        - title
        - address
      properties:
        title:
          type: string
        address:
          type: string

    RequestUpdate:
      type: object
      required:
        - title
        - address
        - status
      properties:
        title:
          type: string
        address:
          type: string
        status:
          type: string